Replace identity dropdown with admin toggle
